+++ CHANGELOG.md
@@ -217,3 +217,4 @@
 - 2023-12-19: Pool leads, bettercontact, deal, conf, +3
 - 2023-12-20: Valid conf, campaign, README, discover, +2
 - 2023-12-21: Init mod, main, LICENSE, qualify, +3
+- 2023-12-22: Test Makefile, lib, mod, qualify, +6